No spaces in function names enclosed in reserved word escape symbols
